Add files via upload
stress detection project was developed that uses machine learning algorithms such as Bernoulli Naive Bayes Classifier and data cleaning techniques like STOPWORDS to detect and analyze stress levels in individuals.

The project involves collecting data from a variety of sources, including social media platforms, online forums, and surveys. The data is then cleaned and preprocessed using techniques such as STOPWORDS removal to remove irrelevant and commonly used words. This ensures that the data is relevant and useful for the analysis.
The Bernoulli Naive Bayes Classifier algorithm is used to classify the data into different categories based on the presence or absence of certain keywords and phrases. This allows the project to detect patterns and trends in the data, which can be used to identify potential stress triggers.

The project can also provide recommendations and resources for individuals who are experiencing stress based on their specific needs and circumstances. This can include advice on stress management techniques, resources for mental health support, and information on self-care practices.

To check your own stress levels in your computer follow the following steps to install the data model:
-Firstly download the CSV file(if you want to deal with other dataset please include the dataset in the project environment)
-Run the code in your jupyterNotebook environment "/ML/ML project.ipynb"

And that's it you ae pretty much done.